Make snugget_load and prepare_load into management tasks so you do not have to open a shell to do them
